+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.1
+
+- Added `DepsUsedChecker` ignores own referenced package.
+
 ## 0.3.0
 
 - Added `DepsUsedChecker` implementation.